+++ app_estudo/domain/__init__.py
@@ -3,11 +3,17 @@
 from .listening_evaluation import ListeningEvaluation, evaluate_listening_item
 from .source_media import SourceMedia, SourceMetadata
 from .study_item import CuratedStudyItem
+from .transcript_normalization import NormalizationResult, normalize_transcript_text
+from .transcript import CuratedTranscript, RawTranscript
 
 __all__ = [
 	"CuratedStudyItem",
+	"CuratedTranscript",
 	"ListeningEvaluation",
+	"NormalizationResult",
+	"RawTranscript",
 	"SourceMedia",
 	"SourceMetadata",
 	"evaluate_listening_item",
+	"normalize_transcript_text",
 ]